streams-ai/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
distro==1.9.0
docstring-parser==0.14.1
executing==2.0.1
fastapi==0.119.0
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
msgpack==1.1.2
openai==2.5.0
py-algorand-sdk==2.11.1
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
PyNaCl==1.6.0
pyteal==0.27.0
python-dotenv==1.1.1
PyYAML==6.0.3
semantic-version==2.10.0
sniffio==1.3.1
starlette==0.48.0
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
